+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.8.0] - 2020-05-18
-- Rename gem from fast_jsonapi to fast-jsonapi
+## [unreleased]
+### Changed
+- Rename gem from fast_jsonapi to fast-jsonapi (#84)
+
+### Fixed
+- Relationship#record_type_for does not assign static record type for polymorphic relationships (#83)
 
 ## [1.7.1] - 2020-05-01
 ### Fixed